requirements.txt
paramiko>=3.0.0
aiohttp>=3.9.0
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
geoip2>=4.8.0
PyYAML>=6.0.1
